feat: add function confirmation on button delete in mysqli crud

+++ mysqli-crud/index.php
@@ -26,12 +26,19 @@
             <td><?= $data['tanggal'] ?></td>
             <td><?= $data['kelas'] ?></td>
             <td>
-                <a href="edit-data.php?id=<?= $data['id'] ?>">EDIT</a> | <a href="aksi-hapus.php?id=<?= $data['id'] ?>">HAPUS</a>
+                <a href="edit-data.php?id=<?= $data['id'] ?>">EDIT</a> | <a onclick="confirmDelete(<?= $data['id'] ?>)"" href="aksi-hapus.php?id=<?= $data['id'] ?>">HAPUS</a>
             </td>
         </tr>
         <?php  
         }
         ?>
     </table>
+
+    <script>
+        function confirmDelete(id) {
+            alert('Apakah anda yakin ingin menghapus data ini?');
+            alert('Data akan dihapus secara permanen');
+        }
+    </script>
 </body>
 </html>
